First pass fud2 integration
